Worked on widget bug
fixed the widget loading
